Updated Model

+++ src/main/java/com/football/assistant/repository/NewsPostRepository.java
@@ -3,5 +3,5 @@ package com.football.assistant.repository;
 import com.football.assistant.domain.NewsPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsPostRepository extends JpaRepository<NewsPost,Long> {
+public interface NewsPostRepository extends JpaRepository<NewsPost,Integer> {
 }
